+++ optd/src/memo/memory.rs
@@ -1,4 +1,6 @@
-use super::*;
+
+
+/*use super::*;
 use crate::cir::*;
 use async_recursion::async_recursion;
 use std::collections::{HashMap, HashSet, VecDeque};
@@ -756,3 +758,4 @@ impl MemoryMemo {
         }
     }
 }
+*/
